requirements.txt
click==6.6
django-environ==0.4.0
geopy==1.11.0
httplib2==0.9.2
overpass==0.4.0
Pillow==2.6.1
pytest==2.9.2
redis==2.10.5
rq==0.6.0
